# openapi/aisi.v1.yaml
openapi: 3.0.3
info:
  title: AISI / Dakwah Depok API
  version: 1.0.0
  description: |
    Kontrak SSOT API AISI (Go).
    Envelope sukses: `{ success: true, data, message?, pagination? }`
    Envelope error: `{ success: false, message }`
servers:
  - url: /api/v1
paths:
...
  /points/me:
    get:
      operationId: myPoints
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OK
  /points/leaderboard:
    get:
      operationId: pointsLeaderboard
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OK
  /points/manual:
    post:
      operationId: manualPoints
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OK
  /notifications:
    get:
      operationId: listNotifications
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OK
  /notifications/{id}/read:
    put:
      operationId: readNotification
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/id'
      responses:
        '200':
          description: OK
  /notifications/read-all:
    put:
      operationId: readAllNotifications
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OK
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    id:
      name: id
      in: path
      required: true
      schema:
        type: string
...